# vecta/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vecta Platform API
  version: 1.0.0
  description: |
    Life-as-a-Service for F-1 international students.
    
    **Authentication:** All `/api/v1/*` endpoints (except webhooks) require
    `Authorization: Bearer <VectaIDToken>` (RS256 JWT).
    
    **F-1 Compliance:** `POST /api/v1/mobility/rides/accept` and
    `POST /api/v1/mobility/driver/go-online` always return `403 F1_VISA_COMPLIANCE_VIOLATION`
    regardless of token — these routes are architectural dead ends.
  contact:
    email: api@vecta.io
  license:
    name: Proprietary
    url: https://vecta.io/terms
servers:
  - url: https://api.vecta.io
    description: Production
  - url: https://api.staging.vecta.io
    description: Staging
  - url: http://localhost:4000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    VectaIDToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT issued by Vecta identity service. Contains zero PII.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:   { type: string, example: INVALID_TOKEN }
        message: { type: string }
...
    MaskedBalance:
      type: object
      description: Balance shown as a range — exact figure never exposed
      properties:
        tier:        { type: string, enum: [LOW, MEDIUM, HIGH, VERY_HIGH] }
        rangeLabel:  { type: string, example: '$10,000 – $25,000' }
        lastUpdated: { type: string, format: date-time }
...
paths:
...
  /api/v1/identity/verify/initiate:
    post:
      summary: Initiate Didit NFC passport scan session
      security: [{ VectaIDToken: [] }]
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId]
              properties:
                studentId: { type: string, format: uuid }
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  sdkToken:  { type: string }
                  qrCodeUrl: { type: string, format: uri }
                  expiresAt: { type: string, format: date-time }
  /api/v1/identity/token/mint:
    post:
      summary: Mint a Vecta ID token (post-KYC)
      security: [{ VectaIDToken: [] }]
      tags: [Identity]
      responses:
        '201':
          description: Token minted and registered for single-use
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: RS256 JWT }
        '403':
          description: KYC not approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/identity/token/verify:
    post:
      summary: Verify a Vecta ID token (landlord use — single-use consumption)
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:             { type: string }
                landlordIp:        { type: string }
                landlordUserAgent: { type: string }
                landlordEmail:     { type: string, format: email }
      responses:
        '200':
          description: Token valid and consumed. Tiered view returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  view: { type: object, description: Tiered student view }
                  tier: { type: string, enum: [ANONYMOUS, VERIFIED, TRUSTED] }
        '401':
          description: Token expired or invalid
        '409':
          description: Token already used (single-use enforcement)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:     { type: string, enum: [TOKEN_ALREADY_USED] }
                  usedAt:    { type: string, format: date-time }
                  usedByIp:  { type: string }
...
  /api/v1/identity/banking/balance:
    get:
      summary: Get masked balance range (never exact balance)
      security: [{ VectaIDToken: [] }]
      tags: [Banking]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaskedBalance' }

  # --- Housing ---
...
  /webhooks/didit:
    post:
      summary: Didit NFC verification webhook
      tags: [Webhooks]
      security: []
      responses:
        '200': { description: Accepted }
        '401': { description: HMAC verification failed }
...